requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
cerebras-cloud-sdk>=1.0.0
python-dotenv>=1.0.0
